Fixat test till FileDownloader. Även en klass exception som vi ska använda senare

+++ src/FileDownloader.java
@@ -3,26 +3,27 @@ import java.io.*;
 import java.net.*;
 
 public class FileDownloader {
-	private BufferedInputStream bis;
-	private FileOutputStream fos;
-	private BufferedOutputStream bos;
-
-	public FileDownloader(String url, String filename)
-			throws MalformedURLException, IOException {
-		bis = new BufferedInputStream(new URL(url).openStream());
-		fos = new FileOutputStream(filename);
-		bos = new BufferedOutputStream(fos, 1024);
-
-
-	}
 	
-	public void WriteToFile() throws IOException{
-		byte[] buf = new byte[4 * 1024]; // 4K buffer
-		int bytesRead;
-		while ((bytesRead = bis.read(buf)) != -1) {
-			bos.write(buf, 0, bytesRead);
+	public static void DownloadFile(String url, String filename) throws BotMException {
+		File file = new File(filename);
+		if(file.exists()){
+			throw new BotMException("The file already exists");
 		}
-		bos.close();
-		bis.close();
+		try {
+			BufferedInputStream bis = new BufferedInputStream(new URL(url).openStream());
+			FileOutputStream fos = new FileOutputStream(filename);
+			BufferedOutputStream  bos = new BufferedOutputStream(fos, 1024);
+			byte[] buf = new byte[4 * 1024];
+			int bytesRead;
+			while ((bytesRead = bis.read(buf)) != -1) {
+				bos.write(buf, 0, bytesRead);
+			}
+			bos.close();
+			bis.close();
+		} catch (MalformedURLException e) {			
+			throw new BotMException("The url is wrong");
+		} catch (IOException e) {			
+			throw new BotMException("I/O problem");
+		}
 	}
 }
